+++ src/com/intellij/idea/plugin/hybris/common/HybrisConstants.kt
@@ -163,6 +163,7 @@ object HybrisConstants {
     const val JAR_MODELS = "models.jar"
     const val SETTINGS_DIRECTORY = ".settings"
     const val EXTERNAL_TOOL_BUILDERS_DIRECTORY = ".externalToolBuilders"
+    const val SPOCK_META_INF_SERVICES_DIRECTORY = "resources/META-INF/services"
     const val WEB_ROOT_DIRECTORY = "webroot"
     const val ACCELERATOR_ADDON_DIRECTORY = "acceleratoraddon"